requirements.txt
Cython ==0.29.21
dataclasses==0.7
decorator==4.4.2
h5py==3.1.0
importlib-metadata==2.0.0
importlib-resources== 3.3.0
json5==0.9.5
jsonpatch==1.26
jsonpointer==2.0
jsonschema==3.2.0
MarkupSafe==1.1.1
numpy == 1.19.4
opencv-python==4.4.0.46
packaging==20.4
paddlenlp== 2.2.0
paddlepaddle-gpu==2.2.1
pandas==1.1.4
pandocfilters       1.4.3
Pillow ==8.0.1
plotly == 4.12.0
pytz==2020.4
requests==2.25.0
scikit-learn== 0.23.2
scipy==1.5.4
six== 1.15.0
tqdm==4.62.3

